Rename all go-lookslike/lookslike refs to go-lookslike
