+++ isdef/int.go
@@ -3,8 +3,8 @@ package isdef
 import (
 	"fmt"
 
-	"github.com/elastic/go-lookslike/lookslike/llpath"
-	"github.com/elastic/go-lookslike/lookslike/llresult"
+	"github.com/elastic/go-lookslike/llpath"
+	"github.com/elastic/go-lookslike/llresult"
 )
 
 func intGtChecker(than int) ValueValidator {
