+++ results_test.go
@@ -20,8 +20,8 @@ package lookslike
 import (
 	"testing"
 
-	"github.com/elastic/go-lookslike/lookslike/llpath"
-	"github.com/elastic/go-lookslike/lookslike/llresult"
+	"github.com/elastic/go-lookslike/llpath"
+	"github.com/elastic/go-lookslike/llresult"
 
 	"github.com/stretchr/testify/assert"
 )
