+++ validator/validator.go
@@ -1,6 +1,6 @@
 package validator
 
-import "github.com/elastic/go-lookslike/lookslike/llresult"
+import "github.com/elastic/go-lookslike/llresult"
 
 // Validator is the result of Compile and is run against the map you'd like to test.
 type Validator func(interface{}) *llresult.Results
